requirements.txt
crewai
litellm
gitpython
markdown
pdfkit
langchain_community
langchain-openai